comentando a classe calculadora

+++ src/main/java/org/example/Calculadora.java
@@ -2,20 +2,24 @@ package org.example;
 
 public class Calculadora {
 
-    public double somaDoisNumerosInteiros(double n1, double n2) {
+    // Método 1: Recebe dois números inteiros e realiza a soma deles
+    public double somaDoisNumerosInteiros(int n1, int n2) {
         return n1 + n2;
     }
 
-    public double subtraiDoisNumerosInteiros(double n1, double n2) {
+    // Método 2: Recebe dois números inteiros e realiza a subtração deles
+    public double subtraiDoisNumerosInteiros(int n1, int n2) {
         return n1 - n2;
     }
 
-    public double multiplicaDoisNumerosInteiros(double n1, double n2) {
+    // Método 3: Recebe dois números inteiros e realiza a multiplicação deles
+    public double multiplicaDoisNumerosInteiros(int n1, int n2) {
         return n1 * n2;
     }
 
-    public double divideDoisNumerosInteiros(double n1, double n2) {
-        return n1 / n2;
+    // Método 4: Recebe dois números inteiros e realiza a divisão deles
+    public double divideDoisNumerosInteiros(int n1, int n2) {
+        return (double) n1 / n2;
     }
 
 }
